requirements.txt
pandas
numpy
scikit-learn
seaborn
matplotlib
scipy
joblib
streamlit
pytest
pyyaml
plotly
